refactor(cli)!: move CLI to arca.flow.cli namespace package
Restructure the CLI as a peer of arca.flow.core under the shared arca.*
namespace, so the orchestrator wheel and the CLI wheel can be shipped
and installed independently while contributing to the same import root.

- cli/arca_flow_cli/ -> cli/arca/flow/cli/ (regular package nested under
  the arca and arca.flow namespace packages)
- internal imports updated from arca_flow_cli.X to arca.flow.cli.X
  across the CLI, its tests, and the cli/ docs
- dev.py target paths updated: CORE_TARGETS now points at arca/flow/core,
  CLI_TARGETS at cli/arca
- env.py clean() walks cli/arca instead of arca_flow_cli
- _LOGO resource loader now addresses the package by its new import name
- cli/pyproject.toml: script entry points and af alias map to
  arca.flow.cli.app:main; hatch wheel build declares packages = ["arca"]
  so the CLI wheel ships only cli/arca/** without an arca/__init__.py
- uv.lock refreshed to reflect arca-flow-core manifest rename

BREAKING CHANGE: external consumers importing arca_flow_cli as a Python
module must update to arca.flow.cli. The arca-flow and af binaries,
their command surface, flags, logo, and env vars are unchanged.
