+++ cli/README.md
@@ -48,8 +48,8 @@ direnv allow  # one-time approval
 | `arca-flow check [--scope ...]` | Run all quality checks (ruff, mypy, pytest) |
 
 The `--scope` flag controls which code is checked:
-- `core` (default): `arca_flow`, `arca_flow_tests`
-- `cli`: `cli/arca_flow_cli`, `cli/tests`
+- `core` (default): `arca/flow/core`, `arca_flow_tests`
+- `cli`: `cli/arca`, `cli/tests`
 - `all`: both core and CLI
 
 ### Environment
