+++ cli/pyproject.toml
@@ -14,5 +14,8 @@ dependencies = [
 ]
 
 [project.scripts]
-arca-flow = "arca_flow_cli.app:main"
-af = "arca_flow_cli.app:main"
+arca-flow = "arca.flow.cli.app:main"
+af = "arca.flow.cli.app:main"
+
+[tool.hatch.build.targets.wheel]
+packages = ["arca"]
